+++ lib/examples/bss_data.rs
@@ -1,7 +1,7 @@
 use std::error::Error;
 
 fn main() -> Result<(), Box<dyn Error>> {
-    let interface = kawaiifi::default_interface().expect("Expected to find a wireless interface");
+    let interface = kawaiifi::default_interface()?.expect("Expected to find a wireless interface");
     let scan = interface.scan_blocking()?;
     print_bss_data(&scan);